Basic support for projects with only pom.xml
If there's a project.clj everything is as before, but if not we not look
for a pom.xml as well.

With the current feature set, this should be fine.
The only dependency on cljr--project-dir are cljr--project-file and
cljr--project-files where it's used to find the project root dir.  This
usage should still be fine.

The only function that actually uses the project file for anything is
cljr--depends-on-p.  At present this function simply does a string
search in the project file to match the dependency.  This usage should
work just as well for the artifacts in pom.xml files.

Closes #27 until somebody who doesn't use Leiningen shows up.