Add cljr-sort-project-dependencies
Calling this fn will sort all the dependency vectors in project.clj.

When cljr-add-project-dependency gets merged in it makes sense to extend
this with a cljr-auto-sort-project-dependencies variable, to keep the
vectors sorted.

If the dependency vector contains any comments, these will bubble to the
top.  It would be possible to associate one, or more comments, with the
following dependency, but--shiiiiiiiiiiit.  Just use end-of-line
comments in project.clj!