Rely on cider for var-info when searching/renaming
I've updated the required cider version, from 0.6 to 0.8-cvs and added
it as a development dependency.  There was already a `(require 'cider)`
in clj-refactor.el so this won't really affect anything.

I've removed the branching on cider version, it was getting to be
annoying to maintain and it's not something I want to keep around
long-term while cider is below 1.0.

This commit also adds keybindings for find-symbol and rename-symbol and
corrects previous spelling mistakes (middlewere instead of midleware).