Add cljr-add-missing-libspec
`cljr-add-missing-libspec` will try to resolve the symbol at point and
require or import the missing var.  If there are more than one var match
the user is prompted. If there's only one result it is added without
user intervention.  If the symbol at point is of the form
`edn/read-string` the resulting libspec will be aliased to `edn`.

This depends on the middleware to do its thing.