Update cljr-extract-function to match refactor-nrepl
We're now sending the name of the ns down to the middleware for
analysis.  Any unbound vars in the ns will become function parameters.  This
will generally do the right thing when we remove all forms following the
newly created function prior to running our analysis.

One limitation of this approach is that shadowing locals won't be used
in the newly created function.  E.g if you re-use the name `str` locally
we won't find `str` as an unbound because a function already has that
name in the global namespace.  Shadowing clojure.core in this manner is
considered bad practice anyway, so I'm not going to spend any energy on
solving this in the near future.  Another argument in favor of writing a
lisp-2, I suppose.

This commit also fixes (extracted-fn ) having a trailing space at the new
callsite when no arguments were passed.