Fix add-missing-libspec with records and deftypes
This also fixes an assumption that anything with capitalized names were
classes, which bit me when using prismatic/schema.

The middleware had to be changed to return more information.  We have to
know whether what we're importing is a missing ns, a missing
deftype/defrecord or a missing class.  It now returns an alist of the
following form ((candidate1 . type1)  (candidate2 . type2)..) where
:type is :record, :type, :class or :ns.  When we're importing interfaces
a class is generated and the import mechanism the same so no effort is
made to distinguish between tehse cases.

The case of missing types and records is most complex because we have to
both require the ns, in order to trigger compilation, and then import
the resulting class.