[Fix #35] cljr-rename-file => cljr-rename-file-or-dir
This relies on the middleware to do the actual moving.  The client is
only responsible for indenting the ns form, of each affected file,
because the middleware lacks the ability to pretty print the ns with
proper indentation.  This is something I hope to remedy later, but for
now this is good enough.

The execution of this op is currently wrapped in a y-or-n-prompt but I'm
wondering if it shouldn't include a stern warning message as well.  I've
done nothing in the middleware to recover from fatal errors, so if this
goes wrong the project will be left in a terrible state.  I thought
about admonishing the user to create a temporary commit, as a restore
point in the event of catastrophic failure, but thought better of it as
it would be annoying after seeing it the first time and I'd feel
compelled to add a defcustom to disable the warning.