Fix promote-fn getting confused by nearby literals
On promoting an fn to a literal `clj--goto-fn-definition` would favor
going to the nearest fn literal, instead of the nearest fn or fn literal.

This caused us to start working on a fn literal further up in the toplevel
form instead of the newly created fn when such a literal was present.