[Fix #200] extract-function and locals
The middleware expects the point to be inside the sexp for which we
want to find unbound locals.  When `extract-function` was run when
point on the opening paren the locals wouldn't be found correctly.

We now first move to the enclosing sexp and then increment point
before calling the middleware, so we're always inside the right sexp.

This probably fixes another bug where only the inner-most locals would
be found and extracted because we were calling the middleware for
locals before moving to the right position.