[Fix #203] Fail earlier when failing in `rs`
The user isn't prompted for a new name until we've queried the
middleware for occurrences of the symbol to be renamed.  This keeps us
from wasting the user's time by collecting information we can't act
on.

The previous solution was slightly better in the sense that when emacs
'freezes' it's clear that it's happening because the user's request is
being handled.  There's now a noticeable hang before we prompt the
user for the new name, as we wait for the middleware to update any
ASTs and return all the occurrences of the symbol to rename.

I've tried to remedy this situation by putting a message in the
minibuffer.  That solution isn't perfect either, because sometimes the
op is so fast (cache hit on all ASTs) that you barely have time to
register that a message flew by.