[#Fix 217] Parameterise auto-ns test declarations
- Bunch of new defcustoms: `cljr-midje-test-declaration`,
  `clj-expectations-test-declarations`,
  `cljr-cljc-clojure-test-declarations` and
  `cljr-clojure-test-declarations`.
- Default is no longer `:refer :all` anywhere.
- Require `expectations :as e`
- Require `clojure.test :as t`
- Require `midje.sweet :as midje`

`:refer :all` is a generally frowned upon now and it's not available in
cljs so in a world of cljc these are likely to be less and less used
now.

If the user is unhappy with our choice of aliases for their test
namespaces they can override them using the new defcustoms.