[Fix #215] Add middleware op for magic requires
- Without middleware everything works as before.
- `cljr-magic-require-namespaces` is now more of a seed value and if the
  middleware is active only plays a role in new projects.
- On ambiguous aliases we prompt for resolution.  Hopefully this will drive
  users to pick unique aliases which improves readability.
- We keep a cache of aliases around, to keep things snappy.  On larger
  projects the aliases are probably fairly stable.
- The cache is updated,async, on repl init and everytime `clean-ns` is called.

If the caching proves to be annoying we can change this easily to be a
sync request for fresh data. If *that* proves too slow we can also
easily cache the alias data in the middleware so we don't gather alias
data from files that are unchanged since last scan.