[Fix #224] Don't remove nil values in create-msg
When an even number of nils were removed we sent nonsense down to the
middleware.  When an odd number of nils were removed the response
couldn't be turned into a map and caused the entire repl the crash in a
way I didn't think possible.

I've no idea why I thought this was a good idea in the first place.

I'm now aiming to write more code under the influence so I can use that
as a defense when I encounter weird code like this in the future.