[Fix #225] Don't request var-info on locals
The problem was caused when multiple candidates were returned from
`cider-var-info` and the user had to take action to disambiguate.

We mitigate the chance of this happening by first checking if the var is
locally defined.  If so it makes no sense to ask `cider-var-info` for
information it doesn't have.

I've only ever seen this prompt for class members (where CIDER is pretty
clueless so it takes a shotgun approach) so I don't expect we'll ever
have to deal with multiple candidates again since inlining of methods or
classes is unsupported.