Refactor improve name, simplify code
In `fe` we were passing around `sexp-forms`, but we never used the first
arguments in the receiving functions.  If we drop the first argument,
then we're left with `args`, a list representing the arguments at the
example site. The code dealing with `args` can be a bit easier to reason
about because the list is short (less `cddaddr` etc :p).