[Fix #238] Boot support for `ap`
The syntax is identical and apparently `hd` also works (probably by
accident and we're modifying the real classpath instead of the intended
faux-classpath, but this seems better than nothing) so this fix was
trivial.

I can't think of any real issues involved in modifying the real
classpath, instead of the faux-classpath, in terms of the current repl
session, so until someone files a bug I'm not going to worry about this
difference between leiningen and boot.