[Fix #269] Document clobbering CIDER keybinding
I don't want to change our recommended keybinding to something which is
unbound because CIDER has taken up all the prime keyboard real estate
already.

Because clj-refactor is used a lot--hopefully!--when writing clojure
code it should be given a keybinding which is both fast and easy on the
hands.  `C-c C-m` fits the bill and has the additional benefit that it
isn't taken by something very important.