Removal of s.el as a dependency (#372)
* Replaced s-join with string-join

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el with subr-x.el

All the `thread-last' got unwound due to the re-write as s.el's `s-join'
expected separator first then the strings. Whereas subr-x.el's
`string-join' expects strings and then the separator.

* Replaced s-trim with string-trim

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el with subr-x.el

* Replaced s-blank? with string-blank-p

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el with subr-x.el

* Replaced s-concat with concat

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el with subr-x.el

* Replace s-starts-with-p with string-prefix-p

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replace s-(suf|pre)fix-p with string-(suf|pre)fix-p

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replaces s-equals-p with string-equal

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replaced s-ends-with-p with string-suffix-p

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replaced s-replace

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replaced s-matches-p

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* cljr--project-dir returns blank string instead of nil

When using cljr--project-dir with the stock Emacs function
string-remove-prefix, like this:

  (string-remove-prefix (cljr--project-dir) path)

instead of s.el's s-chop-prefix:

  (s-chop-prefix (cljr--project-dir) path)

it'll bomb out because cljr--project-dir will return nil, like it does
when running the tests.

But if it returns a blank string, then there is no issue.

* Replaced s-split with split-string

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Remove s-chop-(pre|suf)fix with string-remove-(pre|suf)fix

* Removed s-(pre|ap)pend

* Switch to downcase from s-downcase

As part of this issue:

https://github.com/clojure-emacs/clj-refactor.el/issues/293

* Remove s-present?

As part of the removal of s.el as per
https://github.com/clojure-emacs/clj-refactor.el/issues/293 using
functions that come as part of Emacs

* Replaced s-contains-p

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replace s-match

As per https://github.com/clojure-emacs/clj-refactor.el/issues/293
replacing s.el functions with the ones that come as default with Emacs

* Replace s-dashed-words

As part of the drive to replace the dependency on the functions from
s.el replacing s-dashed-words with function that relies only on
functions that come as part of Emacs release.

* Replaced s-with

thread-last seems to be doing the same thing and is part of Emacs'
subr-x.el whereas s-with is part of s.el

* Replace s-slice-at

As part of the attempt to remove the dependency on s.el, replaced
s-slice-at with a simpler/cruder function.

* Less parens

Removed the unnecessary parenthesis within thread-last macros

* Removed s.el as a dependency

As per the issue
https://github.com/clojure-emacs/clj-refactor.el/issues/293 removed s.el
so that there is one less dependency for this mode that isn't part of
the standard Emacs release.

* Remove s-lines

Replacing it with split-string, but with the parameter taken from
s-lines

* Created cljr--string-present-p

As part of the extraction of the dependencies on s.el, it was replaced
with Emacs native functions, but then it was deemed that their usage was
not straightforward enough, hence cljr--string-present-p

* Cleaning up the intent

Onliner instead of if statement

* Fix for the omission with split-string

It should have been a thread first since the switch from s.el's s-split
takes regexp as the first argument and the string as the last. Whereas
split-string takes the arguments the other way around.

* Made cljr--whitespacep null safe

After removing s-blank? from s.el, just using string-blank-p wasn't null
safe.

Added the null check on Benedek Fazekas' prompting.

* Let diet

No need for the let clause, removing it

* Reused cljr--string-present-p

Using it in a few more places

* Avoiding repetition

Used mapcar in order to avoid the repetition