Let user specify default language context
When the user enters an apparent namespace like `somelib/`, and
the language context is ambiguous, clj-refactor creates a popup
asking the user to specify whether the context is "clj" or "cljs".
This change allows the user to specify a language context which
should be assumed, so that clj-refactor will choose that context
in ambiguous cases rather than creating a popup. Default behavior
is still to create the popup; this change only creates a variable
that lets the user override that behavior.
Tested fairly extensively at REPL; it didn't seem substantial
enough to warrant a new automated test (especially since I've
never worked with Cucumber).