clean-ns: send both the absolute and relative path to the server
This makes use of a change to refactor-nrepl [1], but should be safe when
using older versions of refactor-nrepl, as in that case it simply uses the
absolute path, which is generally preferred over the relative one.

When used with the change to refactor-nrepl it will try first the absolute, then
the relative path, and use the first that can be found.

[1]: https://github.com/clojure-emacs/refactor-nrepl/pull/251