Send nREPL-based paths instead of CIDER-based paths
When `cider-path-translations` is in use (e.g., when CIDER is running
on the host, but the nREPL is running in a VM or a Docker container)
we need to translate the local CIDER-based paths to the remote
nREPL-based paths for those operations that send paths to nREPL.

We use the new `cider-to-nrepl-filename-function` that is introduced
in CIDER pull request #2897, that should be applied before this one.

Also includes a fix for #442.