[FIX #470] Choose deps.edn over pom.xml as project file
https://github.com/clojure-emacs/clj-refactor.el/issues/470
Causes a bug when trying to add a dependency
when you have`deps.edn` and `pom.xml` files in you project.

This happens due to selecting the `pom.xml` as the project file
here:
https://github.com/clojure-emacs/clj-refactor.el/blob/6db85b37b57497b56d97d5e5512160e5db85f798/clj-refactor.el#L2262

But then not finding the `deps` in the next line.
The code then assumes we are using a `leiningen` based project.