Add a step to upgrade cask
This is a (potential) fix for the tests failing against Emacs 25. I
used this config file as a template to create the config file for
vedang/pdf-tools@6ddfda8b and faced the same problem. The failure was
in Cask not being able to install `cl-lib`, which seemed strange to
me. I guessed that the emacs-25 image might not have been baked
properly.

Adding this step fixed the problem for my repository.