Complete requires matching an alias prefix, grouping by cljs or cljc
After triggering a require with cljr-slash, calculate a list of all completions
from the middleware. This is returned in the form of a nested hash of type =>
alias-name => namespace to require. However, many cljs, clj namespaces can
overlap if referencing cljc, so filter the namespaces into the form of a list
of (alias, namespace, types), where types is the set of source file types the
require has been observed in. Then further reduce the completion list to only
those namespaces with an alias equal to the string prior to the / from
cljr-slash. Finally, execute a completing-read on the list of options, and use
that to magically require a namespace.

Lots of efficiency problems here, as computing the whole index is likely
expensive. However, it might be valuable to detect if there is a known option or
change the prompt on an unrecognized namespace.