Allow alias or namespace to match before slash
This should handle adding a include if another file is referring to the
namespace by it's full name. As example, typing:

    (cljs.pprint/)

Should autoinsert a require for cljs.pprint using whichever alias the project
previously used. However, cljs.pprint will still be the name before the slash,
even though the alias is inserted.