Decomplect finding valid aliases from candidate namespace generation
This should be the exact same behavior as previously. However originally the first
`when-let` in `cljr-slash` was calculating if the slash was a potential
namespace alias, what that alias was, and then converting into a list of
candidate `aliases`. Some of that logic is directly in `cljr-slash` and some of
it was in `cljr--magic-requires-lookup-alias.`

After this change, `cljr-slash` detects if it's a potential alias, then
calculates the alias at point, filtering out any existing references using
`cljr--unresolved-alias-ref`. If there is an alias at point, it passes the
it on to `cljr--magic-require-lookup-alias`, which then
combines that information with either alias to namespace mappings from the
middleware or from the `cljr-magic-require-namespaces.`