Use cljr--resolve-alias to double check before inserting require
Attempted to add tests using `cljr--in-namespace-declaration`, but that method
appears to always return nil, and isn't actually working. Left that be on the
other non `cljr-magic-require-prompts-include-context` path to avoid changes in
behavior.

Given current behavior this test is not strictly necessary on the prompting
path, as candidates with existing references should already be filtered out.
Leaving it for now to keep the behavior close to before. Also, there is the
possibility that the `cljr--magic-prompt-or-select-namespace` could allow users
to type in a custom libspec at the prompt, in which case we would still need to
filter to ensure it was not a duplicate.

Added some test coverage to `cljr--resolve-alias`.