Change `cljr-ns-alias-at-point` to detect symbols at point before /
Addresses https://github.com/clojure-emacs/clj-refactor.el/issues/524.

Prevously ns aliases could include quasiquote, sharp-quote and was not
identifying dotted aliases. After reading through the docs on valid namespace
alias, it appears to be any valid symbol. I tested briefly, and all of the
following namespace aliases are valid in both Clojure and Clojurescript.

foo.bar or foo.bar.baz
test? or ?test or a?b
a:b (a::b is legal in cljs only)
tl'an or name'
l33t or a1
foo$bar

I *suspect* that clojure-mode would benefit from providing a robust
clojure-symbol-at-point which could be re-used here. It has a robust symbol
regexp in `clojure--sym-regexp`, which could then be trimmed of leading : and
such, but that is internal to clojure-mode, so I thought it better if it could
be exposed there before using it here.