Add `cljr-slash-uses-suggest-libspec` flag to prefer new middleware
If the `cljr-slash-uses-suggest-libspec` defcustom is enabled, the libspec
recommendations will come from the `suggest-libspec` middleware instead of
`namespace-aliases` by way of the `cljr--prompt-or-select-libspec` method.

No longer filtering the resulting libspec to check if
cljr--in-namespace-declaration is true, as that should be handled by the
backend, or if the user types in their own completing-read value, presumably
they want to include that regardless of the existing namespaces.

Completing-read is allowed to return user input, regardless of whether that is
a known libspec.