Introduce `cljr--reader-conditional-context` (#541)
Addresses problems identified in
https://github.com/clojure-emacs/clj-refactor.el/issues/533.

Previously, `cljr--goto-reader-conditional` and
`cljr--point-in-reader-conditional-branch-p` appeared to have problems with
infinite loops as they looked for specific language values in the reader
conditional.

`cljr--reader-conditional-context` starts at the beginning of the closest
enclosing reader-conditional, and walks through each language, context pair
returning the language name if the point is somewhere between the start of the
language keyword or the next language pair. It will also bail and return nil if
it searches beyond the scope of the reader-conditional or there is no language
symbol it can identify.

`cljr--language-context-at-point` has been amended to use the
`cljr--reader-conditional-context` to return both the language of the file & the
current context to search from.