Expand cljr magic require namespaces documentation (#555)
* alias `edn` to clojure.edn

* Add minimal documentation of `:only` keyword

* limit `io` to `clj`

* backport `:only` keyword support for `cljr-magic-require-namespaces`

This is handled by the suggest-libspec middleware automatically, so should not
be user visible on latest versions. However, if cljr-slash-uses-suggest-libspec
is disabled, the alist format was broken with an :only keyword when parsed by
cljr--magic-requires-lookup-alias. This fixes that lookup and extracts the
libspec from the first element of the list if it's not a dotted pair.